Use 32 bit integers

+++ _cpp/_cpp/main.cpp
@@ -5,15 +5,13 @@
 //  Created by Evren Sen on 2024-10-29.
 //
 
-#include <iostream>
-#include <memory>
 #include <atomic>
 #include <thread>
 
-std::atomic<size_t> counter;
+std::atomic<int32_t> counter;
 
 void add() {
-    std::size_t iterations = 10000000000;
+    std::int64_t iterations = 10000000000;
 
     while (iterations != 0) {
         counter.fetch_add(1, std::memory_order_relaxed);
@@ -22,7 +20,7 @@ void add() {
 }
 
 void sub() {
-    std::size_t iterations = 10000000000;
+    std::int64_t iterations = 10000000000;
 
     while (iterations != 0) {
         counter.fetch_sub(1, std::memory_order_relaxed);
